churn/requirements.txt
numpy
pandas
matplotlib
tensorflow
scikit-learn
onnx 
tf2onnx 
torch
torchvision
transformers
accelerate
python-dotenv
flask
flask_cors
onnxruntime
imbalanced-learn
datasets 
accelerate 
peft 
bitsandbytes
jupyter
nbconvert
lightgbm